services/api-gateway/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlmodel==0.0.14
asyncpg==0.29.0
psycopg2-binary==2.9.9
pydantic==2.5.3
python-multipart==0.0.6

pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
sqlalchemy-utils==0.41.1
